Move finalization registry to its own directory
